Fix for 'out of range'
git-svn-id: http://svn.l2jfree.com/l2jfree@6756 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-datapack/sql/items.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS `items` (
   `custom_type1` INT DEFAULT 0,
   `custom_type2` INT DEFAULT 0,
   `mana_left` SMALLINT NOT NULL DEFAULT -1,    -- duration for shadow items
-  `time` INT(13) NOT NULL DEFAULT 0,            -- duration for time limited items
+  `time` DECIMAL(13) NOT NULL DEFAULT 0,       -- duration for time limited items
   `attributes` varchar(50) NOT NULL DEFAULT '',
   PRIMARY KEY (`object_id`),
   KEY `key_owner_id` (`owner_id`),
